remove unnecessary comments
