+++ src/App.css
@@ -64,3 +64,9 @@ body {
   margin: 15px;
   /* display: inline-block; */
 }
+
+.center {
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+}
