+++ src/screens/PatientDetails.jsx
@@ -25,7 +25,6 @@ const PatientDetails = () => {
     const search = patients?.filter((ele) => {
       return ele.patientId === id;
     });
-    console.log(search);
     setPatientDetails(search);
   }, [patients, id]);
 
@@ -47,7 +46,7 @@ const PatientDetails = () => {
             <p>
               Date of Registration: <b> {ele.date}</b>
             </p>
-            <img src={ele.imageUrl} alt="file" width={200} />
+            <img className="center" src={ele.imageUrl} alt="file" width={200} />
           </div>
         );
       })}
