+++ src/screens/PatientsList.jsx
@@ -47,16 +47,15 @@ const PatientsList = () => {
             <p>
               Date of Registration: <b> {patient.date}</b>
             </p>
+            <img
+              className="center"
+              src={patient.imageUrl}
+              alt="file"
+              width={200}
+            />
           </div>
         );
       })}
-
-      {/* <Button
-        title="Search Patient"
-        onClick={() => {
-          navigate("/search");
-        }} 
-      /> */}
     </div>
   );
 };
